MSVC 64 bit compile warnings fix

+++ src/yajl_parser.h
@@ -72,7 +72,7 @@ yajl_render_error_string(yajl_handle hand, const unsigned char * jsonText,
 /* A little built in integer parsing routine with the same semantics as strtol
  * that's unaffected by LOCALE. */
 long long
-yajl_parse_integer(const unsigned char *number, unsigned int length);
+yajl_parse_integer(const unsigned char *number, size_t length);
 
 
 #endif
